requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.0
async-timeout==4.0.3
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
build==1.3.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
chromadb==1.4.0
click==8.1.8
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.1
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.45.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.2.3
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.81
langchain-google-genai==2.1.12
langchain-text-splitters==0.3.11
langsmith==0.4.37
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.14.0
numpy==2.0.2
oauthlib==3.3.1
onnxruntime==1.19.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
posthog==5.4.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pypdf==6.5.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soupsieve==2.8.1
SQLAlchemy==2.0.45
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
tornado==6.5.4
tqdm==4.67.1
typer==0.21.0
typer-slim==0.21.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.3.0
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
